test: refactoring profile collections and caching #4

+++ app/src/main/java/com/meninocoiso/bscm/data/remote/dto/user/SimplifiedUser.kt
@@ -13,6 +13,6 @@ data class SimplifiedUser(
     val username: String,
     val avatarUrl: String? = null,
     val bannerUrl: String? = null,
-    val accentColor: Int? = null,
+    val accentColor: Long? = null,
     val isVerified: Boolean? = false
 ) : Parcelable
